Updated AggregatedDataPointV4 to use doubles instead of decimals. 4.5M

+++ 1BillionRowChallenge/Models/AggregatedDataPointV4.cs
@@ -2,11 +2,11 @@
 
 public class AggregatedDataPointV4
 {
-    public decimal Min;
+    public double Min;
 
-    public decimal Max;
+    public double Max;
 
-    public decimal Sum;
+    public double Sum;
 
-    public decimal AmountOfDataPoints;
+    public double AmountOfDataPoints;
 }
